+++ test/cases/option-js-inline-keepAttributes/src/app1.js
@@ -1,0 +1,1 @@
+console.log('keep attributes: id, type');